pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["wjsgus95 <wjsgus95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.102.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
